feat(admin): restructure admin pages and refactor components
- Moved and renamed admin-related components and pages for better organization.
- Created dedicated pages for AdminAnnouncementPage, ConfigPage, and VisibilityPage.
- Updated references to use the new page locations in the app.
- Removed the old EmptyAnnouncementForm component as part of the restructuring.
